WSL half works!
Translates paths on the way in but not on the way out.

The right thing to do is to probably add a config option (LSP) for wsl path translation. Maybe with a prefix option to match wsl.conf.
And also to maybe break out the logic needed for wsl <-> windows path translation into a separate, extensively-tested crate.